pyproject.toml
[project]
name = "realtime-api-async-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "websockets>=13.1",
    "python-dotenv>=1.0.1",
    "openai>=1.51.0",
    "pyaudio>=0.2.14",
    "numpy>=2.1.1",
    "pydantic>=2.9.2",
    "pytest>=8.3.3",
    "pytest-asyncio>=0.24.0",
    "pyperclip>=1.9.0",
    "firecrawl-py>=1.3.0",
    "pillow>=10.4.0",
    "rich>=13.9.2",
    "pandas>=2.2.3",
    "psycopg2-binary>=2.9.9",
    "duckdb>=1.1.2",
    "matplotlib>=3.9.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"



[project.scripts]
main = "realtime_api_async_python.main:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
